+++ core/components/frontbar/lexicon/en/default.inc.php
@@ -9,10 +9,18 @@
 
 // Settings lexicons
 $_lang['setting_frontbar.css_url'] = 'User\'s CSS file';
-$_lang['setting_frontbar.css_url_desc'] = 'URL to user-defined CSS file which overrides the default style sheet e.g. /assets/css/frontbar.min.css. Leave this field blank to use the default setting.';
+$_lang['setting_frontbar.css_url_desc'] = 'Path to user-defined CSS file which overrides the default style sheet e.g. /assets/css/frontbar.min.css. Leave this field blank to use the default setting.';
+$_lang['setting_frontbar.sui_css'] = 'Semantic UI CSS';
+$_lang['setting_frontbar.sui_css_desc'] = 'Path to Semantic UI\'s CSS file used for dropdown menus';
+$_lang['setting_frontbar.sui_css_trans'] = 'Semantic UI Transition CSS';
+$_lang['setting_frontbar.sui_css_trans_desc'] = 'Path to Semantic UI\'s CSS file used for transition effects in dropdown menus';
 $_lang['setting_frontbar.fa_url']  = 'FontAwesome';
-$_lang['setting_frontbar.fa_url_desc']  = 'URL to FontAwesome\'s style sheet. Leave blank to use the default setting.';
+$_lang['setting_frontbar.fa_url_desc']  = 'Path to FontAwesome\'s style sheet. Leave blank to use the default setting.';
 $_lang['setting_frontbar.jq_url']  = 'Path to JQuery';
-$_lang['setting_frontbar.jq_url_desc']  = 'URL to JQuery. Leave blank to use the default setting.';
+$_lang['setting_frontbar.jq_url_desc']  = 'Path to JQuery. Leave blank to use the default setting.';
+$_lang['setting_frontbar.sui_js'] = 'Semantic UI JS';
+$_lang['setting_frontbar.sui_js_desc'] = 'Path to Semantic UI\'s JS file used for dropdown menus';
+$_lang['setting_frontbar.sui_js_trans'] = 'Semantic UI Transition JS';
+$_lang['setting_frontbar.sui_js_trans_desc'] = 'Path to Semantic UI\'s JS file used for transition effects in dropdown menus';
 $_lang['setting_frontbar.js_url']  = 'User\'s JS file';
-$_lang['setting_frontbar.js_url_desc']  = 'URL to user-defined Frontbar\'s Javacript file';
+$_lang['setting_frontbar.js_url_desc']  = 'Path to user-defined Frontbar\'s JS file';